+++ tutor/templates/apps/openedx/settings/lms/production.py
@@ -16,7 +16,4 @@ ALLOWED_HOSTS = [
 # Required to display all courses on start page
 SEARCH_SKIP_ENROLLMENT_START_DATE_FILTERING = True
 
-# Allow insecure oauth2 for local interaction with local containers
-OAUTH_ENFORCE_SECURE = False
-
 {{ patch("openedx-lms-production-settings") }}